+++ infra/provider.tf
@@ -13,3 +13,13 @@ provider "aws" {
     tags = local.tags
   }
 }
+
+provider "aws" {
+  alias   = "usea1"
+  region  = "us-east-1"
+  profile = var.aws_profile
+
+  default_tags {
+    tags = local.tags
+  }
+}